command & control working. new scripts: remote_exec.sh, upload_to_hosts

+++ scripts/run_gateway.sh
@@ -4,10 +4,19 @@
 set -e
 
 if [ $# -lt 1 ]; then
-	echo "usage: $0 <func_config.json> " 
+	echo "usage: $0 <experiment-name> "
 	exit 1
 fi
 
+lhost=$(hostname)
+
+if [ ! -d "$HOME/$1" ]; then
+	echo "[$lhost]: Experiment not loaded"
+	exit 42
+fi
+
+experiment=$1
+cd $HOME/$experiment
 source common.env # fix this
 
 if [ -z "$COMMON_ENV_H" ]; then
@@ -15,12 +24,12 @@ if [ -z "$COMMON_ENV_H" ]; then
     exit 2
 fi
 
-func_config=$1
-
-rm -rf $BASE_DIR/outputs
+[ -d "$BASE_DIR/outputs" ] && rm -rf $BASE_DIR/outputs
 mkdir -p $BASE_DIR/outputs
 
 $NIGHTCORE_ROOT/bin/release/gateway \
-    --func_config_file=$BASE_DIR/$1 \
-    --v=1 2>$BASE_DIR/outputs/gateway.log
+    --func_config_file=$BASE_DIR/func_config.json \
+    --v=1 2>$BASE_DIR/outputs/gateway.log &
+
+echo "[$lhost]: gateway has been deployed for experiment $experiment"
